+++ .eslintrc.js
@@ -16,7 +16,8 @@
  */
 const process = require('process');
 
-const hideErrorsInDev = process.env.HIDE_ERRORS_IN_DEV === 'true';
+//const hideErrorsInDev = process.env.HIDE_ERRORS_IN_DEV === 'true';
+const hideErrorsInDev = true;
 
 const productionError = hideErrorsInDev ? 'warn': 'error';
 const productionWarn = hideErrorsInDev ? 'off': 'warn';
@@ -82,7 +83,7 @@ module.exports = {
     '@typescript-eslint/no-unsafe-call': productionWarn,
     '@typescript-eslint/no-unsafe-return': productionWarn,
     '@typescript-eslint/no-unsafe-member-access': productionWarn,
-    'import/namespace': 'off', // productionError,
+    'import/namespace': productionError,
     'no-console': productionWarn,
     'no-debugger': productionError,
     eqeqeq: 'error',
