+++ .idea/misc.xml
@@ -10,11 +10,6 @@
   <component name="ProjectType">
     <option name="id" value="jpab" />
   </component>
-  <component name="SwUserDefinedSpecifications">
-    <option name="specTypeByUrl">
-      <map />
-    </option>
-  </component>
   <component name="WebPackConfiguration">
     <option name="mode" value="DISABLED" />
   </component>
