+++ .idea/vcs.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="VcsDirectoryMappings">
+    <mapping directory="$PROJECT_DIR$" vcs="Git" />
     <mapping directory="$PROJECT_DIR$/lunatrace/bsl/frontend/src/scss/proprietary-theme" vcs="Git" />
     <mapping directory="$PROJECT_DIR$/tools/license-checker/skywalking-eyes" vcs="Git" />
   </component>
