+++ lunatrace/bsl/frontend/src/components/finding/VulnerabilityTableItem.tsx
@@ -11,15 +11,16 @@
  * limitations under the License.
  *
  */
+import {findingIsNotIgnored} from "@lunatrace/lunatrace-common/build/main";
 import classNames from 'classnames';
 import React, { useState } from 'react';
 import { FloatingLabel, Form, FormControl, OverlayTrigger, Spinner, Tooltip } from 'react-bootstrap';
 import { XSquare } from 'react-feather';
 import { useNavigate, useParams } from 'react-router-dom';
 
-import api from '../../../api';
-import { ConfirmationDailog } from '../../../components/ConfirmationDialog';
-import { CvssInferredWarning } from '../../../components/CvssInferredWarning';
+import api from '../../api';
+import { ConfirmationDailog } from '../ConfirmationDialog';
+import { CvssInferredWarning } from '../CvssInferredWarning';
 
 import { Finding } from './types';
 
@@ -40,6 +41,8 @@ export const VulnerabilityTableItem: React.FC<VulnerabilityTableItemProps> = ({
   const [showConfirmation, setShowConfirmation] = useState(false);
   const [ignoreNote, setIgnoreNote] = useState('');
 
+  const findingIsIgnored = !findingIsNotIgnored(finding);
+
   const ignoreVuln = async () => {
     await insertVulnIgnore({
       objects: [
@@ -53,6 +56,27 @@ export const VulnerabilityTableItem: React.FC<VulnerabilityTableItemProps> = ({
     });
     setIgnoreNote('');
   };
+
+  const getIgnoreColumn = () => {
+    if (findingIsIgnored) {
+      return "ignored"
+    }
+
+    if (insertVulnIgnoreState.isLoading) {
+      return (
+        <Spinner size="sm" animation="border" />
+      );
+    }
+    return (
+      <XSquare
+        className="ignore-vuln-button"
+        onClick={(e) => {
+          e.stopPropagation();
+          setShowConfirmation(true);
+        }}
+      />
+    );
+  }
 
   return (
     <>
@@ -77,19 +101,7 @@ export const VulnerabilityTableItem: React.FC<VulnerabilityTableItemProps> = ({
             <CvssInferredWarning inferred={finding.vulnerability.cvss_inferred || false} placement="top" />{' '}
           </td>
           <td>{finding.fix_versions || 'unknown'}</td>
-          <td>
-            {insertVulnIgnoreState.isUninitialized ? (
-              <XSquare
-                className="ignore-vuln-button"
-                onClick={(e) => {
-                  e.stopPropagation();
-                  setShowConfirmation(true);
-                }}
-              />
-            ) : (
-              <Spinner size="sm" animation="border" />
-            )}
-          </td>
+          <td>{getIgnoreColumn()}</td>
         </tr>
       </OverlayTrigger>
       <ConfirmationDailog
