+++ lunatrace/bsl/frontend/src/components/finding/types.ts
@@ -11,7 +11,7 @@
  * limitations under the License.
  *
  */
-import { GetBuildDetailsQuery } from '../../../api/generated';
+import { GetBuildDetailsQuery } from '../../api/generated';
 
 export type Finding = NonNullable<GetBuildDetailsQuery['builds_by_pk']>['findings'][number];
