+++ lunatrace/bsl/frontend/src/utils/string-utils.ts
@@ -21,3 +21,7 @@ export function truncateString(str: string, maxLength: number) {
   }
   return str;
 }
+
+export function maybePluralize(count: number, noun: string, includeCount = false, suffix = 's') {
+  return (includeCount ? `${count} ` : '') + `${noun}${count !== 1 ? suffix : ''}`;
+}
